flask/requirements.txt
Flask==3.0.3
mysql-connector-python==8.0.33
cryptography==41.0.3
Flask-SQLAlchemy==3.0.3
PyMySQL==1.0.3
Flask-Cors==3.0.10
tenacity==8.1.0


